fix: [P4ADEV-4094] circular dependency
