+++ src/routes/DebtTypeOrgCreate/steps/Step2Behaviour/Step2Behaviour.test.tsx
@@ -7,7 +7,7 @@ import { FormProvider, useForm, FieldValues } from 'react-hook-form';
 import { setOrganizationId } from '../../../../store/OrganizationIdStore';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { step2Schema } from './schema';
-import { PaymentMethodOption } from './components/PaymentMethodSelector';
+import { PaymentMethodOption } from '../../types';
 
 vi.mock('../../hooks/useNotificationConfig', () => ({
   useNotificationConfigurations: () => ({
